Create header file for flyby_ui.c. Add include guards.
