+++ flyby_defines.h
@@ -1,3 +1,8 @@
+#ifndef FLYBY_DEFINES_H_DEFINED
+#define FLYBY_DEFINES_H_DEFINED
+
 #define MAX_NUM_CHARS		1024
 #define MAX_NUM_TRANSPONDERS	10
 #define MAX_NUM_SATS		250
+
+#endif
