+++ flyby_ui.c
@@ -51,46 +51,11 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <predict/predict.h>
+#include "flyby_ui.h"
 
 #define EARTH_RADIUS_KM		6.378137E3		/* WGS 84 Earth radius km */
 #define HALF_DELAY_TIME	5
 #define	KM_TO_MI		0.621371		/* km to miles */
-
-struct tle_db_entry {
-	long satellite_number;
-	char name[MAX_NUM_CHARS];
-	char line1[MAX_NUM_CHARS];
-	char line2[MAX_NUM_CHARS];
-};
-
-struct tle_db {
-	int num_tles;
-	struct tle_db_entry tles[MAX_NUM_SATS];
-	char filename[MAX_NUM_CHARS];
-};
-
-struct sat_db_entry {
-	long satellite_number;
-	bool squintflag;
-	double alat;
-	double alon;
-	int num_transponders;
-	char transponder_name[MAX_NUM_TRANSPONDERS][MAX_NUM_CHARS];
-	double uplink_start[MAX_NUM_TRANSPONDERS];
-	double uplink_end[MAX_NUM_TRANSPONDERS];
-	double downlink_start[MAX_NUM_TRANSPONDERS];
-	double downlink_end[MAX_NUM_TRANSPONDERS];
-	unsigned char dayofweek[MAX_NUM_TRANSPONDERS];
-	int phase_start[MAX_NUM_TRANSPONDERS];
-	int phase_end[MAX_NUM_TRANSPONDERS];
-};
-
-struct transponder_db {
-	int num_sats;
-	struct sat_db_entry sats[MAX_NUM_SATS];
-	char filename[MAX_NUM_CHARS];
-	bool loaded;
-};
 
 double reduce(value,rangeMin,rangeMax)
 double value, rangeMin, rangeMax;
@@ -147,10 +112,6 @@ char   *mStr;
 void bailout(string)
 char *string;
 {
-	/* This function quits ncurses, resets and "beeps"
-	   the terminal, and displays an error message (string)
-	   when we need to bail out of the program in a hurry. */
-
 	beep();
 	curs_set(1);
 	bkgdset(COLOR_PAIR(1));
@@ -208,13 +169,6 @@ int n;
 char KepCheck(line1,line2)
 char *line1, *line2;
 {
-	/* This function scans line 1 and line 2 of a NASA 2-Line element
-	   set and returns a 1 if the element set appears to be valid or
-	   a 0 if it does not.  If the data survives this torture test,
-	   it's a pretty safe bet we're looking at a valid 2-line
-	   element set and not just some random text that might pass
-	   as orbital data based on a simple checksum calculation alone. */
-
 	int x;
 	unsigned sum1, sum2;
 
@@ -396,11 +350,6 @@ void SaveTLE(struct tle_db *tle_db)
 
 int AutoUpdate(char *string, struct tle_db *_tle_db, predict_orbital_elements_t **orbits)
 {
-	/* This function updates PREDICT's orbital datafile from a NASA
-	   2-line element file either through a menu (interactive mode)
-	   or via the command line.  string==filename of 2-line element
-	   set if this function is invoked via the command line. */
-
 	int num_sats = _tle_db->num_tles;
 	struct tle_db_entry *tle_db = _tle_db->tles;
 
@@ -853,12 +802,8 @@ double GetStartTime(const char* info_str)
 	return ((double)DayNum(mm,dd,yy)+((hr/24.0)+(min/1440.0)+(sec/86400.0)));
 }
 
-int Print(title,string,mode)
-char *title, *string, mode;
+int Print(const char *title, const char *string, char mode)
 {
-	/* This function buffers and displays orbital predictions
-	   and allows screens to be saved to a disk file. */
-
 	char type[20], head2[81];
 	int key, ans=0;
 	static char buffer[5000], lines, quit;
@@ -958,17 +903,8 @@ char *title, *string, mode;
 	return (quit);
 }
 
-int PrintVisible(title,string)
-char *title, *string;
+int PrintVisible(const char *title, const char *string)
 {
-	/* This function acts as a filter to display passes that could
-	   possibly be visible to the ground station.  It works by first
-	   buffering prediction data generated by the Predict() function
-	   and then checking it to see if at least a part of the pass
-	   is visible.  If it is, then the buffered prediction data
-	   is sent to the Print() function so it can be displayed
-	   to the user and optionally logged to a file. */
-
 	static char buffer[10000];
 	char line[80], plus, asterisk, visible;
 	int x, y, quit=0;
@@ -1023,11 +959,6 @@ char *title, *string;
 
 void Predict(const char *name, predict_orbital_elements_t *orbital_elements, predict_observer_t *qth, char mode)
 {
-	/* This function predicts satellite passes.  It displays
-	   output through the Print() function if mode=='p' (show
-	   all passes), or through the PrintVisible() function if
-	   mode=='v' (show only visible passes). */
-
 	bool should_quit = false;
 	bool should_break = false;
 	char data_string[MAX_NUM_CHARS];
@@ -1141,8 +1072,6 @@ void Predict(const char *name, predict_orbital_elements_t *orbital_elements, pre
 		refresh();
 	}
 }
-
-enum celestial_object{PREDICT_SUN, PREDICT_MOON};
 
 void celestial_predict(enum celestial_object object, predict_observer_t *qth, predict_julian_date_t time, struct predict_observation *obs) {
 	switch (object) {
@@ -1276,9 +1205,6 @@ bool KbEdit(int x, int y, int num_characters, char *output_string)
 
 void ShowOrbitData(int num_orbits, struct tle_db_entry *tle_db, predict_orbital_elements_t **orbital_elements_array)
 {
-	/* This function permits displays a satellite's orbital
-	   data.  The age of the satellite data is also provided. */
-
 	int c, x, namelength, age;
 	double an_period, no_period, sma, c1, e2, satepoch;
 	char days[5];
@@ -1441,9 +1367,6 @@ void QthEdit(const char *qthfile, predict_observer_t *qth)
 
 void SingleTrack(bool once_per_second, double horizon, int orbit_ind, int num_orbits, predict_orbital_elements_t **orbital_elements_array, predict_observer_t *qth, struct sat_db_entry *sat_db_entries, struct tle_db_entry *tle_db_entries, rotctld_info_t *rotctld, rigctld_info_t *downlink_info, rigctld_info_t *uplink_info)
 {
-	/* This function tracks a single satellite in real-time
-	   until 'Q' or ESC is pressed. */
-
 	int     ans;
 	bool	downlink_update=true, uplink_update=true, readfreq=false;
 
@@ -1948,11 +1871,6 @@ double scrk, scel;
 
 void MultiTrack(predict_observer_t *qth, int num_orbits, predict_orbital_elements_t **orbital_elements_array, struct tle_db_entry *tle_db, char multitype, char disttype)
 {
-	/* This function tracks all satellites in the program's
-	   database simultaneously until 'Q' or ESC is pressed.
-	   Satellites in range are HIGHLIGHTED.  Coordinates
-	   for the Sun and Moon are also displayed. */
-
 	int 		*satindex = (int*)malloc(sizeof(int)*num_orbits);
 
 	double		*aos = (double*)malloc(sizeof(double)*num_orbits);
